Add memory64 support to wasm-smith
This commit adds support to wasm-smith to generate modules that use
memory64, namely 64-bit memories and all indices passed to memory-using
operations use 64-bit indexes instead of 32-bit indicies. The main
changes to support this were in the `code_builder.rs` file around
generating memory indices to work with and such.

This commit also refactors the fuzzers a bit as well as `SwarmConfig`.
The fuzzers now enable all features in `SwarmConfig`, even those that
are disabled by default in `SwarmConfig` due to not being implemented in
many locations (such as `memory64`). The goal here is that the fuzzing
of this repository specifically should fuzz everything supported by
wasm-smith, not just the default items.

+++ crates/wasm-encoder/src/code.rs
@@ -150,7 +150,7 @@ impl Function {
 #[derive(Clone, Copy, Debug)]
 pub struct MemArg {
     /// A static offset to add to the instruction's dynamic address operand.
-    pub offset: u32,
+    pub offset: u64,
     /// The expected alignment of the instruction's dynamic address operand
     /// (expressed the exponent of a power of two).
     pub align: u32,
@@ -162,10 +162,10 @@ impl MemArg {
     fn encode(&self, bytes: &mut Vec<u8>) {
         if self.memory_index == 0 {
             bytes.extend(encoders::u32(self.align));
-            bytes.extend(encoders::u32(self.offset));
+            bytes.extend(encoders::u64(self.offset));
         } else {
             bytes.extend(encoders::u32(self.align | (1 << 6)));
-            bytes.extend(encoders::u32(self.offset));
+            bytes.extend(encoders::u64(self.offset));
             bytes.extend(encoders::u32(self.memory_index));
         }
     }
